+++ src/gr/aueb/cf/cf9/ch18/bankapp/dto/AccountWithdrawDTO.java
@@ -1,4 +1,6 @@
 package gr.aueb.cf.cf9.ch18.bankapp.dto;
 
-public record AccountWithdrawDTO(String iban, double amount) {
+import java.math.BigDecimal;
+
+public record AccountWithdrawDTO(String iban, BigDecimal amount) {
 }
